requirements.txt
numpy
scipy
matplotlib
progressbar2
skimage
